Fuzion dist builder hack needs upstreaming, Refs CRM-15768, make CiviCRM Drupal module version consistent with other Drupal modules.
Drupal module version strings include the core version as 6.x, 7.x, 8.x.

The sed regex has been stripping this information, which leads to CiviCRM being reported somewhat inconsistently in Drupal's update checks. This change better preserves the core version prefix when building CiviCRM releases.

A fix for this was previously committed to civicrm-drupal, but did not "stick" due to the sed in the release build scripts overwriting that.