Fixed for 4.7 CRM-18259 call pledge & pledgePayment specific functions for geting status options.
The current config is to have a long list of contributions statuses which are shared as an option group
with pledge & pledge_payment. The irrelevant ones are then hacked out.

I believe that the correct way to do this is to split out the option groups into 3 separate ones.

However, there is quite a lot of work in re-pointing the various places that get the statuses from contributions.

So, I decided to limit this step to fixing the code to call the correct buildOptions function to get
the list & moving the hacks into there.

This does fix the original issue as the  status was being hacked out for pledge payment
when it should have been kept for pledge

Conflicts:
	CRM/Pledge/BAO/Pledge.php
	CRM/Pledge/Form/Search.php